Significant design change
Previously, hvw-dhcpd was a real DHCP server serving a pool of pre-configured IP addresses, now it has been modified to act as a DHCP relay with caching. Now we have to test this new behavior in real condition. Of course, this is to be more compliant with Patrick H. Fry and Joshua W. Knight CDRA paper.